src/requirements.txt
python-telegram-bot
futsu
boto3
